requirements.txt
altair==4.2.2
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
clip==0.2.0
colorama==0.4.6
comtypes==1.4.5
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1

MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0


psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
#pypiwin32==223  # Consider removing if not on Windows or not necessary
pyserial==3.5
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
#pywin32==306  # Consider removing if not on Windows
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
scipy==1.14.0
seaborn==0.13.2
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
protobuf==3.20.3

SpeechRecognition==3.10.4
streamlit==1.22.0
sympy==1.13.1
tbb==2021.11.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.71
ultralytics-thop==2.0.0
urllib3==2.2.2
watchdog==4.0.1